+++ packages/eyes-api/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- add `CheckSettingsFluent#shadow` method to allow specifying region selectors for elements in shadow dom
 
 ## 1.0.12 - 2021/8/13